requirements.txt
hydra-core>=1.3.2
omegaconf>=2.3.0
transformer-lens>=2.4.0
saelens>=0.1.6
torch>=2.2.0
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
seaborn>=0.12
tqdm>=4.66
datasets>=2.19
sentencepiece>=0.1.99
einops>=0.7
requests>=2.31
transformers>=4.30.0

